Gate Efficiency & Automated Test Runs
Recording gate efficiency statistics to use as part of the report.
Using the same struct as fitness stats.
Exported in CSV report file.
Statistics calculation also added for gate efficiency score.

Programmable truth table file reader re-implemented.
Used for the automated test runs.

Changes to sim_export function.
Exports a text '.rpt' file, contains the parameters and population.
Exports a CSV file, containing the fitness and efficiency stats.

Edits made to main menu options, and inspection functions.
Mainly cosmetics.

New variable, 'gate' for GA class objects.
For containing the efficiency score. Done once per individual.
> Bool eval is now used to skip gate evaluations only.
> Truth table evaluations should not be skipped,
> as there is still some random chance of a 'fluke'.
Also patched setter functions to prevent possible input bugs.

* Possible Memory Leak Problem:
Encountered more memory leak problems again.
Hard to reproduce, occurs on fringe cases.
Haven't tracked down the root cause.
> Strange phenomenon:
> GA helper function compfit_descend when coded in this style (see diff)
> reduces the likelihood of a segfault.
> Compare this to the better way of coding in compfit_ascend.
Fixed(?) by rebooting and clearing old obj files before compiling.
Problem hasn't surfaced since. Worrisome.

Minor changes to comments and others.

+++ HPS/eval.hpp
@@ -63,11 +63,11 @@ unsigned int eval_seq_insp (void);
 
 /* ========== Efficiency Evaluation Functions ========== */
 
-/* uint32_t eval_efficiency (const uint8_t *const *const grid);
+/* uint16_t eval_efficiency (const uint8_t *const *const grid);
 	Efficiency evaluation metric.
 	Set to evaluate which solution is more gate efficient.
 	Using less number of gate is better.
 */
-uint32_t eval_efficiency (const uint8_t *const *const grid);
+uint16_t eval_efficiency (const uint8_t *const *const grid);
 
 #endif
